requirements-backend.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
openai==1.98.0
google-search-results==2.4.2
pydantic==2.4.2